record check

+++ app/Http/Controllers/API/CertificateController.php
@@ -36,8 +36,13 @@ class CertificateController extends Controller
     public function get($id)
     {
         $record = Model::find($id);
-        $response = ['code' => 200, 'message' => "Fetched $this->model", 'record' => $record];
-        return response($response);
+        if ($record) {
+            $response = ['code' => 200, 'message' => "Fetched $this->model", 'record' => $record];
+        }
+        else {
+            $response = ['code' => 404, 'message' => "$this->model Not Found"];
+        }
+        return response()->json($response);
     }
 
     public function create(Request $request)
@@ -57,7 +62,7 @@ class CertificateController extends Controller
         $record = Model::create($validated);
         $response = ['code' => 200, 'message' => "Created $this->model", 'record' => $record];
 
-        return response($response);
+        return response()->json($response);
     }
 
     public function update(Request $request)
@@ -81,14 +86,21 @@ class CertificateController extends Controller
         $record->update($validated);
         $response = ['code' => 200, 'message' => "Updated $this->model"];
 
-        return response($response);
+        return response()->json($response);
     }
 
     public function delete($id)
     {
         $record = Model::find($id);
-        Storage::disk('s3')->delete("certificates/$record->image");
-        $record->delete();
-        return response(['code' => 200, 'message' => "Deleted $this->model"]);
+        if ($record) {
+            Storage::disk('s3')->delete("certificates/$record->image");
+            $record->delete();
+            $response = ['code' => 200, 'message' => "Deleted $this->model"];
+        }
+        else {
+            $response = ['code' => 404, 'message' => "$this->model Not Found"];
+        }
+
+        return response()->json($response);
     }
 }
