+++ app/Http/Controllers/API/ItemController.php
@@ -19,14 +19,19 @@ class ItemController extends Controller
     {
         $records = Model::all();
         $response = ['code' => 200, 'message' => "Fetched $this->model" . "s", 'records' => $records];
-        return response($response);
+        return response()->json($response);
     }
 
     public function get($id)
     {
         $record = Model::find($id);
-        $response = ['code' => 200, 'message' => "Fetched $this->model", 'record' => $record];
-        return response($response);
+        if ($record) {
+            $response = ['code' => 200, 'message' => "Fetched $this->model", 'record' => $record];
+        }
+        else {
+            $response = ['code' => 404, 'message' => "$this->model Not Found"];
+        }
+        return response()->json($response);
     }
 
     public function create(Request $request)
@@ -47,7 +52,7 @@ class ItemController extends Controller
         $record = Model::create($validated);
         $response = ['code' => 200, 'message' => "Created $this->model", 'record' => $record];
 
-        return response($response);
+        return response()->json($response);
     }
 
     public function update(Request $request)
@@ -72,14 +77,21 @@ class ItemController extends Controller
         $record->update($validated);
         $response = ['code' => 200, 'message' => "Updated $this->model"];
 
-        return response($response);
+        return response()->json($response);
     }
 
     public function delete($id)
     {
         $record = Model::find($id);
-        Storage::disk('s3')->delete("items/$record->image");
-        $record->delete();
-        return response(['code' => 200, 'message' => "Deleted $this->model"]);
+        if ($record) {
+            Storage::disk('s3')->delete("items/$record->image");
+            $record->delete();
+            $response = ['code' => 200, 'message' => "Deleted $this->model"];
+        }
+        else {
+            $response = ['code' => 404, 'message' => "$this->model Not Found"];
+        }
+
+        return response($response);
     }
 }
