+++ app/Http/Controllers/API/ScheduleController.php
@@ -25,8 +25,7 @@ class ScheduleController extends Controller
                 $records = Model::where('user_id', $user_id)->get();
             }
             $response = ['code' => 200, 'message' => "Fetched $this->model" . "s", 'records' => $records];
-        }
-        else {
+        } else {
             $response = ['code' => 401, 'message' => "User Not Authenticated"];
         }
         return response()->json($response);
@@ -35,8 +34,12 @@ class ScheduleController extends Controller
     public function get($id)
     {
         $record = Model::find($id);
-        $response = ['code' => 200, 'message' => "Fetched $this->model", 'record' => $record];
-        return response($response);
+        if ($record) {
+            $response = ['code' => 200, 'message' => "Fetched $this->model", 'record' => $record];
+        } else {
+            $response = ['code' => 404, 'message' => "$this->model Not Found"];
+        }
+        return response()->json($response);
     }
 
     public function create(Request $request)
@@ -57,7 +60,7 @@ class ScheduleController extends Controller
         $record = Model::create($validated);
         $response = ['code' => 200, 'message' => "Created $this->model", 'record' => $record];
 
-        return response($response);
+        return response()->json($response);
     }
 
     public function update(Request $request)
@@ -80,14 +83,19 @@ class ScheduleController extends Controller
         $record->update($validated);
         $response = ['code' => 200, 'message' => "Updated $this->model"];
 
-        return response($response);
+        return response()->json($response);
     }
 
     public function delete($id)
     {
         $record = Model::find($id);
-        $record->delete();
-        return response(['code' => 200, 'message' => "Deleted $this->model"]);
+        if ($record) {
+            $record->delete();
+            $response = ['code' => 200, 'message' => "Deleted $this->model"];
+        } else {
+            $response = ['code' => 404, 'message' => "$this->model Not Found"];
+        }
+        return response()->json($response);
     }
 
     public function changeStatus(Request $request)
@@ -101,6 +109,6 @@ class ScheduleController extends Controller
         $record->update($validated);
         $response = ['code' => 200, 'message' => "Updated Status"];
 
-        return response($response);
+        return response()->json($response);
     }
 }
