+++ app/Http/Controllers/API/TestimonialController.php
@@ -34,8 +34,12 @@ class TestimonialController extends Controller
     public function get($id)
     {
         $record = Model::find($id);
-        $response = ['code' => 200, 'message' => "Fetched $this->model", 'record' => $record];
-        return response($response);
+        if ($record) {
+            $response = ['code' => 200, 'message' => "Fetched $this->model", 'record' => $record];
+        } else {
+            $response = ['code' => 404, 'message' => "$this->model Not Found"];
+        }
+        return response()->json($response);
     }
 
     public function create(Request $request)
@@ -49,7 +53,7 @@ class TestimonialController extends Controller
         $record = Model::create($validated);
         $response = ['code' => 200, 'message' => "Created $this->model", 'record' => $record];
 
-        return response($response);
+        return response()->json($response);
     }
 
     public function update(Request $request)
@@ -65,13 +69,19 @@ class TestimonialController extends Controller
         $record->update($validated);
         $response = ['code' => 200, 'message' => "Updated $this->model"];
 
-        return response($response);
+        return response()->json($response);
     }
 
     public function delete($id)
     {
         $record = Model::find($id);
-        $record->delete();
-        return response(['code' => 200, 'message' => "Deleted $this->model"]);
+        if ($record) {
+            $record->delete();
+            $response = ['code' => 200, 'message' => "Deleted $this->model"];
+        } else {
+            $response = ['code' => 404, 'message' => "$this->model Not Found"];
+        }
+
+        return response()->json($response);
     }
 }
